refactor: 투표 개표 분리
