+++ src/main/java/mafia/mafiatogether/domain/Room.java
@@ -1,8 +1,6 @@
 package mafia.mafiatogether.domain;
 
-import java.util.HashMap;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentHashMap;
 import lombok.AccessLevel;
@@ -19,7 +17,7 @@ import mafia.mafiatogether.domain.job.JobType;
 public class Room {
 
     private final Map<String, Player> players;
-    private final Map<Player, Player> votes;
+    private final Vote vote;
     private Status status;
     private final RoomInfo roomInfo;
     private final Chat chat;
@@ -29,7 +27,7 @@ public class Room {
     public static Room create(final RoomInfo roomInfo) {
         return new Room(
                 new ConcurrentHashMap<>(),
-                new ConcurrentHashMap<>(),
+                Vote.create(),
                 Status.WAIT,
                 roomInfo,
                 Chat.chat(),
@@ -80,38 +78,10 @@ public class Room {
     public void votePlayer(final String name, final String targetName) {
         final Player player = players.get(name);
         final Player target = players.get(targetName);
-        votes.put(player, target);
+        vote.choose(player, target);
     }
 
     public String getVoteResult() {
-        return countVotes().getName();
-    }
-
-    private Player countVotes() {
-        final Map<Player, Integer> voteCounts = new HashMap<>();
-        for (final Player player : votes.values()) {
-            voteCounts.put(player, voteCounts.getOrDefault(player, 0) + 1);
-        }
-        return findMaxVotedPlayer(voteCounts);
-    }
-
-    private Player findMaxVotedPlayer(final Map<Player, Integer> voteCounts) {
-        int maxCount = 0;
-        int count = 0;
-        Player votedPlayer = Player.NONE;
-        for (Entry<Player, Integer> voteCount : voteCounts.entrySet()) {
-            if (maxCount == voteCount.getValue()) {
-                count++;
-            }
-            if (maxCount < voteCount.getValue()) {
-                maxCount = voteCount.getValue();
-                count = 1;
-                votedPlayer = voteCount.getKey();
-            }
-        }
-        if (count > 1) {
-            return Player.NONE;
-        }
-        return votedPlayer;
+        return vote.getVoteResult();
     }
 }
